NEW display controls in a consistent manner
As a work around to embedding the HistoryViewerField within a regular
edit form (such as within a GridFieldDetailForm) allow a developer to
opt-in to hiding the save/publish actions, and pushing the revert button
to the bottom of the CMS screen to keep actions for the current screen
in the CMS consistent. This can be done by adding a class to the Tab
that hosts the history viewer.