NEW Dropdown atop history viewer for holding actions
In order to allow a user to enter or exit compare mode there is a button
on the heading row that will reveal more actions when clicked. At
current the only action present is a toggle to enter or exit compare
mode, and reflect the state (check means compare mode is active).