NEW Adjust history viewer to allow for compare mode
History viewer is the parent of all fields related to comparison mode
viewing, and begins the process by deciding whether to render a detail
view or a list of versions. It now must also factor comparison mode into
this, and so alter the way it handles the list or the detail view.