Logic for Diff rendering moved from Transformation to new Field
Major refactor of Diff Form Transformation
The original approach was implmented as a "get it functional" state, and
a bit of an attempt to circumvent the complexities involved with the
logical order of operations, the nature of a failover (see
ViewableData), and having form fields work in more than one display
facet. However there were limitations in handling more complex data
structures such as when a value from the database was NULL, this would
fail a data comparison check between loaded data and transformation
data. In order to work around this we now simply ignore all these
complexities and pass off to normal errors or exceptions that will be
output with such use. There is now a diff field that will handle the
data explicitly, rather than having the entire transformation happen in
the transformer class. This was originally avoided due to a failover
being an underlay, not an overlay, so only methods that didn't exist on
the superclass would be passed off to the form field we're tryign to
transform, instead of for any method not defined on the class. To
circumvent this issue, we now simply don't (cicumvent this issue), and
rely on users not e.g. transforming a form then attempting to
update one of the old fields.