FIX remove double-margins from version comparison
Previously a fix was inserted to have the content push away from the
outer boundaries of the containing element (eg. the CMS menu), however
this has been resolved in some other commit in history and no longer
presents a problem. Instead it shows a 'double' margin, padding the
detail view with twice as much space as it should have between contained
content, and content outside it's boundaries.